refactor: ImageResolver 로직 분리
